refactor!(support): move opencv related functionality into its own package
