+++ packages/opencv/README.md
@@ -1,0 +1,7 @@
+# @appium/opencv
+
+OpenCV-related helper methods
+
+`npm install @appium/opencv`
+
+This package relies on opencv.js, which is WASMed into a ~7mb dependency and run "natively" in JS. It is therefore standalone without requiring C bindings.
