+++ packages/opencv/test/images/appium-diagram.jpg
