+++ packages/opencv/test/images/cc1.png
