+++ packages/opencv/test/images/cc2.png
