+++ packages/opencv/test/images/cc_rotated.png
