+++ packages/opencv/test/images/findwaldo.jpg
