+++ packages/opencv/test/images/number5.png
