+++ packages/opencv/test/images/waldo.jpg
